chimera: introduce inode generation number
on some workloads we need to detect that file or a directory have
been changed, even if mtime and ctime does not reflect that. This
can happen if time resolution is not good enough or timestams
where reseted back.

To reliably identify inode changes, introduced a new field,
igeneration, which is monotonically incremented on each inode
modification.

RELEASE NOTES:
The first start of namespace will take some time ( ~10-20 min )
for database schema update.


Acked-by: Gerd Behrmann