chimera: do not maintain time-based cached value of FsStat
NFS server by itself can cache the results of fsstat.
JdbcFs sill keeps a cached result, but it updated on demand.
Nevertheless, as this is a time consuming operation,
start an extra thread to perform the update. As long as update
is still in the progress, the last value is returned.
    
The update part takes care that only one thread is started in parallel.