chimera: keep track of tags usage
Motivation:
when directory is deleted, we have to delete all tags references
to it and tags without references have to be deleted.

As tags have attribute set similar to directory objects (as in chimera
they was originally implemented as directories) use nlink count to keep
track or references.

Modification:
use inlink field of t_tags_inodes table to keep track of references. Delete
tags with zero references, when a directory is deleted.

update db changelog to adjust references on the first start.

Result:
postgres unfriendly query replaced with a quicker alternative.

NOTICE: on the first start DB migration may take some minutes. On DB
with 50M entries in t_tags it took < 1 min.